Add: 컴포넌트 합치기
