+++ src/globalStyles.tsx
@@ -2,10 +2,10 @@ import { createGlobalStyle } from 'styled-components';
 import reset from 'styled-reset';
 
 const GlobalStyles = createGlobalStyle`
+  ${reset}
 *{
     box-sizing: border-box;
 }
-    ${reset}
     
 body{
   font-style : normal;
